Add files via upload
The x00 to x14 files are binary chunks which can be reassembled to form a ZIP file which contains the Windows loader program for the XDM12041 firmware which is also included (V2.1.2). To reassemble on a Windows machine you need to use the command line (DOS prompt). Go to the directory where you stored the 15 files and use:

copy/b x* XDM1041.zip